Clean Trash Fill States from Reported Log
Now the Reporter checks if Trash Fill States which have been reported
and are thus stored in the Reported Log have been deleted from the Fill
State Log. If they have been, they are also deleted from the Reported
Log.

This prevents the Reported Log from growing constantly and aligns the
implementation with the specification.